django_back/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.0.0
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.3
daphne==4.1.0
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.11
langchain-community==0.0.26
langchain-core==0.1.53
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.84.0
orjson==3.10.18
packaging==23.2
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyOpenSSL==25.1.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
service-identity==24.2.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
tenacity==8.5.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.0
zope.interface==7.2
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
python-jose==3.3.0
boto3==1.34.162
django-storages==1.14.4
Pillow==10.4.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
